conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
avro-python3=1.10.2=pyhd8ed1ab_0
brotli-python=1.0.9=py312h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.28.1=hd590300_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py312h06a4308_0
cffi=1.16.0=py312h5eee18b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
cryptography=42.0.5=py312hdda0065_1
cyrus-sasl=2.1.27=h54b06d7_7
dnspython=2.4.2=py312h06a4308_0
expat=2.6.2=h6a678d5_0
fastavro=1.9.1=py312h5eee18b_0
flask=2.2.5=py312h06a4308_0
flask-pymongo=2.3.0=pyhd8ed1ab_0
graphql-core=3.2.3=pypi_0
idna=3.4=py312h06a4308_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.3=py312h06a4308_0
keyutils=1.6.1=h166bdaf_0
krb5=1.21.2=h659d440_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=8.7.1=hca28451_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libntlm=1.4=h7f98852_1002
librdkafka=2.3.0=hfe68a65_0
libsqlite=3.45.3=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.2.0=h7e041cc_5
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=hcb278e6_0
markupsafe=2.1.3=py312h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.2.1=hd590300_1
pip=23.3.1=py312h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pymongo=4.6.3=py312h6a678d5_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-confluent-kafka=2.3.0=py312h98912ed_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=4_cp312
readline=8.2=h5eee18b_0
requests=2.31.0=py312h06a4308_1
setuptools=68.2.2=py312h06a4308_0
sqlite=3.41.2=h5eee18b_0
strawberry-graphql=0.227.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
typing-extensions=4.11.0=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py312h06a4308_1
werkzeug=2.3.8=py312h06a4308_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
